pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "secure_credentials"
version = "1.0.0"
description = "A secure credential manager with GUI interface"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
dependencies = [
    "ttkbootstrap>=1.10.1",
    "pykeepass>=4.0.3",
    "pyperclip>=1.8.2",
    "pillow>=9.5.0",
    "pystray>=0.19.4",
    "packaging>=23.1",
    "requests>=2.31.0",
    "olefile>=0.46"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "pyinstaller>=5.13.0"
]

[project.scripts]
secure-credentials = "secure_credentials.src.run_app:main"

[tool.hatch.build.targets.wheel]
packages = ["secure_credentials"] 